# Increase depth and emit a trace when pruning vacuous branches in `Proxy.hs`