Fixes #3973