@geo2a 2024-07-11
No difference in KEVM performance:

```
| Test                                        | georgy-fix-proxy time | master-0932411b9 time | (georgy-fix-proxy/master-0932411b9) time
|---------------------------------------------|-----------------------|-----------------------|------------------------------------------
| mcd/flopper-dent-guy-same-pass-rough-spec.k | 862.93                | 804.53                | 1.072588964985768
| TOTAL                                       | 862.93                | 804.53                | 1.072588964985768
```